A realm of biting cold and hollow darkness
